feat: extract, document and simplify ActiveRetrievalsManager
